backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
pydantic==2.8.2
httpx==0.27.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
scipy==1.13.1
pypdf==4.3.1
nltk==3.8.1
jinja2==3.1.4
bibtexparser==1.4.1
citeproc-py==0.6.0
python-dateutil==2.9.0
